Blank input checker + yes or no checker
Created a function to take a user input to take down their name and return the question to the user if their input was blank. Then created a program to ask the user a yes or no question and return the question if they did not input yes or no.